improved item physics; items are now half the size as before

+++ Assets/Features/Items/CollectableItem.cs
@@ -4,14 +4,14 @@ public class CollectableItem : MonoBehaviour
 {
     [Header("Item Data")]
     [SerializeField] private ItemType itemType;
-    [SerializeField] private int widthInLanes = 1;
-    [SerializeField] private int heightInLanes = 1;
+    [SerializeField] private float widthInLanes = 1f;
+    [SerializeField] private float heightInLanes = 1f;
     [SerializeField] private float mass = 2f;
     [SerializeField] private GameObject stackedPrefab;
 
     public ItemType ItemType => itemType;
-    public int WidthInLanes => widthInLanes;
-    public int HeightInLanes => heightInLanes;
+    public float WidthInLanes => widthInLanes;
+    public float HeightInLanes => heightInLanes;
     public float Mass => mass;
     public GameObject StackedPrefab => stackedPrefab;
     public bool WasCollected { get; private set; }
